chore: renamed the files to follow flutter naming convention
