--- lib/screens/HomePage.dart
@@ -1,66 +1,0 @@
-import 'package:flutter/material.dart';
-
-import '../buttons/RecordingList.dart';
-import '../buttons/headerHalf.dart';
-
-class HomePage extends StatefulWidget {
-  const HomePage({Key? key}) : super(key: key);
-
-  @override
-  State<HomePage> createState() => _HomePageState();
-}
-
-class _HomePageState extends State<HomePage> {
-  // This widget is the root of your application.
-  int _selectedIndex = 0;
-  void _onItemTapped(int index) {
-    setState(() {
-      _selectedIndex = index;
-    });
-  }
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: const Color(0xffF3F7FF),
-      body: Column(
-        children: const [
-          headerHalf(),
-          RecordingList()
-        ],
-      ),
-      bottomNavigationBar: BottomNavigationBar(
-        //backgroundColor: const Color(0xffF3F7FF),
-        items: const <BottomNavigationBarItem>[
-          BottomNavigationBarItem(
-            backgroundColor:  Color(0xffF3F7FF),
-            icon: ImageIcon(
-              AssetImage("assets/images/img_profile.png"),
-            ),
-            label: '',
-          ),
-          BottomNavigationBarItem(
-            icon: ImageIcon(
-              AssetImage("assets/images/img_explore.png"),
-            ),
-            label: '',
-          ),
-          BottomNavigationBarItem(
-            icon: ImageIcon(
-              AssetImage("assets/images/img_recordings.png"),
-            ),
-            label: '',
-          ),
-          BottomNavigationBarItem(
-            icon: ImageIcon(
-              AssetImage("assets/images/img_setting.png"),
-            ),
-            label: '',
-          )
-        ],
-        selectedItemColor:  const Color(0xff3D79FD),
-        unselectedItemColor: Colors.black,
-        onTap: _onItemTapped,
-      ),
-    );
-  }
-}
