+++ lib/screens/home_page.dart
@@ -1,0 +1,51 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+import '../buttons/RecordingList.dart';
+import '../buttons/headerHalf.dart';
+
+class HomePage extends ConsumerWidget {
+  const HomePage({super.key});
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    return Scaffold(
+      backgroundColor: const Color(0xffF3F7FF),
+      body: Column(
+        children: const [headerHalf(), RecordingList()],
+      ),
+      bottomNavigationBar: BottomNavigationBar(
+        //backgroundColor: const Color(0xffF3F7FF),
+        items: const <BottomNavigationBarItem>[
+          BottomNavigationBarItem(
+            backgroundColor: Color(0xffF3F7FF),
+            icon: ImageIcon(
+              AssetImage("assets/images/img_profile.png"),
+            ),
+            label: '',
+          ),
+          BottomNavigationBarItem(
+            icon: ImageIcon(
+              AssetImage("assets/images/img_explore.png"),
+            ),
+            label: '',
+          ),
+          BottomNavigationBarItem(
+            icon: ImageIcon(
+              AssetImage("assets/images/img_recordings.png"),
+            ),
+            label: '',
+          ),
+          BottomNavigationBarItem(
+            icon: ImageIcon(
+              AssetImage("assets/images/img_setting.png"),
+            ),
+            label: '',
+          )
+        ],
+        selectedItemColor: const Color(0xff3D79FD),
+        unselectedItemColor: Colors.black,
+      ),
+    );
+  }
+}
