+++ lib/widgets/alert_dialog_model.dart
@@ -1,0 +1,55 @@
+import 'package:flutter/material.dart';
+
+class DialogUtils {
+  static final DialogUtils _instance = DialogUtils.internal();
+
+  DialogUtils.internal();
+
+  factory DialogUtils() => _instance;
+
+  static void showCustomDialog(BuildContext context,
+      {
+        required String title,
+        String deleteBtnText = "Delete",
+        String saveBtnText = "Save",
+        required Function deleteBtnFunction,
+      }) {
+    showDialog(
+        context: context,
+        builder: (BuildContext context) {
+          return Dialog(
+            shape: RoundedRectangleBorder(
+                borderRadius:
+                BorderRadius.circular(25.0)), //this right here
+            child: Container(
+              height: 200,
+              child: Padding(
+                padding: const EdgeInsets.all(12.0),
+                child: Column(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    const TextField(
+                      decoration: InputDecoration(
+                          border: InputBorder.none,
+                          hintText: 'Name of recording'),
+                    ),
+                    SizedBox(
+                      width: 320.0,
+                      child: TextButton(
+                        onPressed: () {},
+                        //color: const Color(0xFF1BC0C5),
+                        child: const Text(
+                          "Delete",
+                          style: TextStyle(color: Colors.white),
+                        ),
+                      ),
+                    )
+                  ],
+                ),
+              ),
+            ),
+          );
+        });
+  }
+}
