+++ lib/widgets/onboarding_button.dart
@@ -1,0 +1,25 @@
+import 'package:flutter/material.dart';
+
+class OnboardingButton extends StatelessWidget {
+  final String text;
+  final VoidCallback? onPressed;
+
+  const OnboardingButton({
+    super.key,
+    required this.text,
+    this.onPressed,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return ElevatedButton(
+      onPressed: onPressed,
+      style: ElevatedButton.styleFrom(
+        foregroundColor: Colors.black,
+        backgroundColor: const Color(0xffC5D7FE),
+        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(25)),
+      ),
+      child: Text(text),
+    );
+  }
+}
