+++ pubspec.yaml
@@ -36,6 +36,7 @@ dependencies:
   cupertino_icons: ^1.0.2
   audio_waveforms: ^1.0.1
   flutter_riverpod: ^2.3.2
+  riverpod: ^2.3.2
 
 dev_dependencies:
   flutter_test:
